Refactoring Code 1. Separated the creation of handlers from one single function into    separate ones. This makes it easier to handler each conversation individually instead. I also moved the functions to create these handlers into its own file (bot_handlers.py) to declutter the run file.
2. The previous implementation of conversations did not implement
   conversation states well. It used only a single state for each
conversation and how the user progressed through the conversation was
hard to debug. Each conversation now has multiple states to indicate
which stage of the conversation the user is at. An example of this is
updating laundry, where the user will go through conversation states in
this order, LAUN_START -> LAUN_UPDATE -> LAUN_CONF

3. The previous implementation of ending conversations was hard to
   debug, it just returns the ACTION_START state which is the starting
menu that users see on first /start. Now at the end of each
conversation, ConversationHandler.END is called instead and
context.user_data is cleared.

4. Premature operation cancel in each conversation was not handled
   correctly. A new route is now present in each conversation which
enables the user to type '/cancel' to cleanly exit a conversation to
start a new one.

5. User state in a conversation was not being tracked/handled. Now when
   a user starts a conversation, a variable in their context user_data
is updated to indicate the user is busy in a conversation. Attempts to
start a new conversation while in the middle of one will be met with an
error message indicating that the user has to either complete the entire
conversation or cancel premature using /cancel.

6. A new function that searches for usernames containing a specific
   string is created in acl_methods.py. Will update this to a fuzzy
finder to make the process of inputing the name of the user to register
easier.

7. Added more comments into the code for clarity

Notes: Loan conversation is currently not working and has to be
debugged and refactored. Unittests for acl_methods will be done by the
next commit, this mainly focuses on testing functions that manipulate
user data in the database.